requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
cachetools==5.5.1
ccxt==4.4.51
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
discord-py==2.4.0
exceptiongroup==1.2.2
fonttools==4.55.6
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==0.6.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.0
mplfinance==0.12.10b0
multidict==6.1.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycares==4.5.0
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2024.2
requests==2.32.3
rsa==4.9
schedule==1.2.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
ta==0.11.0
tqdm==4.67.1
typing-extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.2
yarl==1.18.3
